+++ packages/html-reporter/src/expandable.tsx
@@ -1,0 +1,46 @@
+/**
+ * Copyright (c) Microsoft Corporation.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+import React from 'react';
+import * as icons from './icons';
+
+import './expandable.css';
+
+export const Expandable: React.FunctionComponent<{
+  summary: React.ReactNode,
+  children: React.ReactNode,
+  className?: string,
+  style?: React.CSSProperties,
+}> = ({ summary, children, className, style }) => {
+  const [expanded, setExpanded] = React.useState(false);
+
+  const handleToggle = (e: React.SyntheticEvent<HTMLDetailsElement>) => {
+    setExpanded(e.currentTarget.open);
+  };
+
+  return (
+    <details
+      style={style}
+      className={className}
+      onToggle={handleToggle}
+    >
+      <summary className='expandable-summary'>
+        {expanded ? icons.downArrow() : icons.rightArrow()}
+        {summary}
+      </summary>
+      {children}
+    </details>
+  );
+};
